Feature: insert mirror local database in cloud

+++ src/services/truck.js
@@ -1,9 +1,10 @@
 const Truck = require('../models/Truck');
+const TruckCloud = require('../models-cloud/Truck');
 
 class TruckService {
   static async getAllTruck() {
     try {
-      const truck = await Truck.findAll({order: [['placa', 'ASC']]});
+      const truck = await Truck.findAll({ order: [['placa', 'ASC']] });
       return truck;
     } catch (error) {
       return error;
@@ -13,8 +14,8 @@ class TruckService {
     try {
       const truck = await Truck.findOne({ where: { "placa": placa } });
       if (!truck) {
-        const newTruck = await Truck.create({ placa: placa });
-        return newTruck;
+        const truck = await Truck.create({ placa: placa, cloud: false });
+        return truck;
       }
       return { 'msg': 'Este caminhão já existe!' };
     } catch (error) {
@@ -27,8 +28,10 @@ class TruckService {
       const truck = await Truck.findOne({ where: { "id": req.body.id } });
       if (truck) {
         await Truck.update({ placa: req.body.placa }, { where: { "id": req.body.id } });
-        const retorno = await Truck.findOne({ where: { "id": req.body.id } });
-        return retorno;
+        const truck = await Truck.findOne({ where: { "id": req.body.id } });
+        return truck;
+      } else {
+        return { 'msg': 'Este caminhão não existe!' };
       }
     } catch (error) {
       return error;
@@ -37,12 +40,47 @@ class TruckService {
 
   static async deleteTruck(placa) {
     try {
-      const truck = await Truck.destroy({ where: { "placa": placa } });
-      const retorno = await Truck.findOne({ where: { "placa": placa } });
-      if (!retorno) {
-        return { "deletado": true };
+      const truck = await Truck.findOne({ where: { "placa": placa } });
+      if (truck) {
+        await Truck.destroy({ where: { "placa": placa } });
+        const retorno = await Truck.findOne({ where: { "placa": placa } });
+        if (!retorno) {
+          return { 'msg': 'Caminhão removido!' };
+        } else {
+          return { 'msg': 'Caminhão não removido!' };
+        }
+      } else {
+        return { 'msg': 'Este caminhão não existe!' };
       }
-      return { "deletado": false };
+    } catch (error) {
+      return error;
+    }
+  }
+
+
+  static async getAll() {
+    try {
+      const truck = await Truck.findAll({ raw: true, where: { "cloud": false } });
+      return truck;
+    } catch (error) {
+      return error;
+    }
+  }
+
+  static async send(trucks) {
+    try {
+      const truck = await TruckCloud.bulkCreate(trucks);
+      return truck;
+    } catch (error) {
+      return error;
+    }
+  }
+
+  static async modify(trucks) {
+    try {
+      for (let i = 0; i < trucks.length; i++) {
+        await Truck.update({ cloud: true }, { where: { "id": trucks[i].id } });
+      }
     } catch (error) {
       return error;
     }
